pyproject.toml
[tool.poetry]
name = "labels"
version = "0.1.0"
description = ""
authors = ["Vic Putz <vbputz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pylabels = "^1.2.1"
reportlab = "^3.6.1"
toolz = "^0.11.1"
click = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
